requirements.txt
black==25.9.0
celery==5.5.3
channels==4.3.1
channels_redis==4.3.0
cinema==0.0.11
daphne==4.2.1
Django==5.2.7
django-celery-beat==2.8.1
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-standardized-errors==0.15.0
drf-yasg==1.21.11
flower==2.0.1
gunicorn==23.0.0
pre-commit==4.4.0
psycopg2-binary==2.9.10
pytest==8.4.2
python-dotenv==1.1.1
python-json-logger==4.0.0
pytz==2025.2
redis==7.0.0
tzdata==2025.2
url-normalize==2.2.1
